requirements.txt
pycryptodome===3.16.0
boto3===1.28.51
sqlalchemy===2.0.21
psycopg2===2.9.7
pytest===7.4.2